requirements.txt
fastapi
uvicorn
asyncpg
sqlalchemy[asyncio]
pydantic
python-dotenv
google-generativeai
